Implement separate source and target replication proxies
Previously if a proxy was specified it was used for both source and target
traffic. However, as mentioned in #2272, since both source and target now are
URLs, instead of one being a "local" database, it makes sense to allow separate
proxy settings for source and target endpoints.

We are still allowing old style, single proxy setting, however if users set
both the old style proxy and a per-endpoint one, an exception is raised about
the settings being mutually exclusive.

Fixes #2272